pyproject.toml
[tool.poetry]
name = "langchain-course"
version = "0.1.0"
description = ""
authors = ["Ahmad Gill <ahmad.gill@adnocdistribution.ae>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
